feat: put notebook events in component
